GRF-240: Move files after discussion
