+++ make-file/user-management.mk
@@ -62,4 +62,4 @@ user-management-coupling-back:
 
 .PHONY: user-management-integration-back
 user-management-integration-back: #Doc: launch PHPUnit integration tests for user-management bounded context
-	APP_ENV=test $(PHP_RUN) vendor/bin/phpunit --testsuite PIM_Integration_Test --filter UserManagement $(F)
+	APP_ENV=test $(PHP_RUN) vendor/bin/phpunit --testsuite PIM_Integration_Test --filter UserManagement $(O)
