+++ src/Akeneo/UserManagement/back/Application/Handler/ListUserGroupHandler.php
@@ -2,12 +2,13 @@
 
 namespace Akeneo\UserManagement\Application\Handler;
 
+use Akeneo\UserManagement\ServiceApi\UserGroup\ListUserGroupHandlerInterface;
 use Akeneo\UserManagement\Domain\Model\Group as DomainGroup;
 use Akeneo\UserManagement\Domain\Storage\FindUserGroups;
-use Akeneo\UserManagement\Infrastructure\ServiceApi\UserGroup\ListUserGroupQuery;
-use Akeneo\UserManagement\Infrastructure\ServiceApi\UserGroup\UserGroup;
+use Akeneo\UserManagement\ServiceApi\UserGroup\ListUserGroupQuery;
+use Akeneo\UserManagement\ServiceApi\UserGroup\UserGroup;
 
-class ListUserGroupHandler
+class ListUserGroupHandler implements ListUserGroupHandlerInterface
 {
     public function __construct(
         private FindUserGroups $findUserGroups
@@ -20,7 +21,6 @@ class ListUserGroupHandler
     public function __invoke(ListUserGroupQuery $query): array
     {
         // @todo implement optional arguments in $query to allow research and pagination
-        // @todo exclude the default user group
         $result = ($this->findUserGroups)(
             $query->getSearchName(),
             $query->getSearchAfterId(),
