+++ src/Akeneo/UserManagement/back/ServiceApi/UserGroup/ListUserGroupHandlerInterface.php
@@ -1,0 +1,22 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Akeneo PIM Enterprise Edition.
+ *
+ * (c) 2022 Akeneo SAS (https://www.akeneo.com)
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Akeneo\UserManagement\ServiceApi\UserGroup;
+
+interface ListUserGroupHandlerInterface
+{
+    /**
+     * @return UserGroup[]
+     */
+    public function __invoke(ListUserGroupQuery $query): array;
+}
